nav_delete

+++ app/controllers/Posts.php
@@ -131,6 +131,12 @@
 
         public function delete($id){
             if($_SERVER['REQUEST_METHOD'] == 'POST'){
+                //Get existing post from model
+                $post = $this->postModel->getPostById($id);
+                // Check for owner
+                if($post->userId != $_SESSION['user_id']){
+                    redirect('posts');
+                }
                 if($this->postModel->deletePost($id)){
                     flash('post_message', 'Post Removed');
                     redirect('posts');
